feat: CDS-1091
