+++ packages/react-compass/src/table/table.stories.tsx
@@ -5,18 +5,22 @@ import {
   faClose,
   faDashboard,
   faFileLines,
+  faHeart,
+  faHeartBroken,
   faPencil,
   faTrashAlt,
+  faUser,
 } from '@fortawesome/free-solid-svg-icons'
-import { Meta } from '@storybook/react'
-import React, { Key, MouseEvent, TouchEvent, useState } from 'react'
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
+import {Meta} from '@storybook/react'
+import React, {Key, MouseEvent, TouchEvent, useState} from 'react'
 import Table, {
   TableCellContext,
   TableColumnDef,
+  TableColumnFiltersState,
   TableHeaderContext,
   TableSortingState,
   useEditableCellContext,
-  TableColumnFiltersState,
 } from '.'
 import {
   Button,
@@ -28,25 +32,26 @@ import {
   Spinner,
   TextField,
 } from '..'
-import StatusComponent from './for story/person-status'
 import {
   LimitRequestStatus,
   Person,
   makeData,
   makeRequestStatusData,
 } from './makeData'
-import { OptionType } from './table'
 import ProgressPercentage from './table-progress'
+import {OptionType} from './types'
+
 export const Basic: React.FC = () => {
   const [page, setPage] = useState(1)
-  const [data, setData] = useState<Person[]>(() => makeData(10))
+  const [data, setData] = useState<Person[]>(() => makeData(100))
   const options: OptionType<Person> = {
     enableSorting: true,
     enableMultiSort: true,
     columnResizeMode: 'onChange',
     manualSorting: false,
     manualFiltering: true,
-    enableRowSelection: (row) => row.original.age > 30,
+    getRowId: (row) => row.id,
+    enableRowSelection: (row) => row.original.age > 20,
   }
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function
@@ -65,7 +70,7 @@ export const Basic: React.FC = () => {
     () => [
       {
         id: 'select',
-        header: ({ table }) => {
+        header: ({table}) => {
           return (
             <div
               style={{
@@ -82,7 +87,7 @@ export const Basic: React.FC = () => {
             </div>
           )
         },
-        cell: ({ row }) => (
+        cell: ({row}) => (
           <div
             style={{
               display: 'flex',
@@ -101,7 +106,7 @@ export const Basic: React.FC = () => {
       },
       {
         id: 'name',
-        header: () => <div style={{ textAlign: 'center' }}>Name</div>,
+        header: () => <div style={{textAlign: 'center'}}>Name</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -144,7 +149,7 @@ export const Basic: React.FC = () => {
       },
       {
         id: 'otherInfo',
-        header: () => <div style={{ textAlign: 'center' }}>Other info</div>,
+        header: () => <div style={{textAlign: 'center'}}>Other info</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -184,7 +189,8 @@ export const Basic: React.FC = () => {
   )
   return (
     <Table
-      data={data}
+      data={data.slice((page - 1) * 10, page * 10)}
+      // data={data}
       columns={columns}
       options={options}
       onManualSorting={onSorting}
@@ -256,7 +262,7 @@ export const Basic: React.FC = () => {
 }
 
 export const ExpandableRow: React.FC = () => {
-  const [data] = React.useState(() => makeRequestStatusData(10))
+  const [data] = React.useState(() => makeRequestStatusData(100))
   const [page, setPage] = useState(1)
   const options: OptionType<LimitRequestStatus> = {
     enableSorting: false,
@@ -270,7 +276,7 @@ export const ExpandableRow: React.FC = () => {
       {
         id: 'expander',
         size: 40,
-        header: ({ table }: TableHeaderContext<LimitRequestStatus, unknown>) => (
+        header: ({table}: TableHeaderContext<LimitRequestStatus, unknown>) => (
           <span
             onClick={table.getToggleAllRowsExpandedHandler()}
             style={{
@@ -288,7 +294,7 @@ export const ExpandableRow: React.FC = () => {
             )}
           </span>
         ),
-        cell: ({ row }: TableCellContext<LimitRequestStatus, unknown>) => (
+        cell: ({row}: TableCellContext<LimitRequestStatus, unknown>) => (
           <span
             onClick={() => row.toggleExpanded()}
             style={{
@@ -310,7 +316,7 @@ export const ExpandableRow: React.FC = () => {
       {
         id: 'accountName',
         accessorKey: 'accountName',
-        header: () => <div style={{ textAlign: 'center' }}>Account Name</div>,
+        header: () => <div style={{textAlign: 'center'}}>Account Name</div>,
         footer: (props: TableHeaderContext<LimitRequestStatus, unknown>) =>
           props.column.id,
       },
@@ -318,7 +324,7 @@ export const ExpandableRow: React.FC = () => {
         id: 'code',
         accessorKey: 'code',
 
-        header: () => <div style={{ textAlign: 'center' }}>Code</div>,
+        header: () => <div style={{textAlign: 'center'}}>Code</div>,
         footer: (props: TableHeaderContext<LimitRequestStatus, unknown>) =>
           props.column.id,
       },
@@ -327,7 +333,7 @@ export const ExpandableRow: React.FC = () => {
         accessorKey: 'requestLimit',
 
         header: () => (
-          <div style={{ textAlign: 'center' }}>New Request Limit</div>
+          <div style={{textAlign: 'center'}}>New Request Limit</div>
         ),
         footer: (props: TableHeaderContext<LimitRequestStatus, unknown>) =>
           props.column.id,
@@ -337,7 +343,7 @@ export const ExpandableRow: React.FC = () => {
         size: 280,
 
         accessorKey: 'status',
-        header: () => <div style={{ textAlign: 'center' }}>Status</div>,
+        header: () => <div style={{textAlign: 'center'}}>Status</div>,
         footer: (props: TableHeaderContext<LimitRequestStatus, unknown>) =>
           props.column.id,
       },
@@ -361,8 +367,8 @@ export const ExpandableRow: React.FC = () => {
         {rowData.progress?.map((processItem) => {
           return (
             <>
-              <p style={{ fontSize: '12px' }}>{processItem.remarks}</p>
-              <p style={{ fontSize: '11px', fontWeight: 200, color: '#878787' }}>
+              <p style={{fontSize: '12px'}}>{processItem.remarks}</p>
+              <p style={{fontSize: '11px', fontWeight: 200, color: '#878787'}}>
                 {processItem.createdAt.toDateString()}
               </p>
               {/* <Divider /> */}
@@ -394,8 +400,8 @@ export const ExpandableRow: React.FC = () => {
             justifyContent: 'space-between',
           }}
         >
-          <div style={{ display: 'flex', alignItems: 'center' }}>
-            <span style={{ width: 250 }}>Rows per page: </span>
+          <div style={{display: 'flex', alignItems: 'center'}}>
+            <span style={{width: 250}}>Rows per page: </span>
             <Dropdown.Select defaultSelectedKey='10'>
               <Dropdown.Item key='10'>10 rows</Dropdown.Item>
               <Dropdown.Item key='20'>20 rows</Dropdown.Item>
@@ -438,8 +444,8 @@ export const EmptyState: React.FC = () => {
     columnResizeMode: 'onChange',
     manualSorting: false,
     initialSortBy: [
-      { id: 'firstName', desc: true },
-      { id: 'lastName', desc: false },
+      {id: 'firstName', desc: true},
+      {id: 'lastName', desc: false},
     ],
   }
 
@@ -447,7 +453,7 @@ export const EmptyState: React.FC = () => {
     () => [
       {
         id: 'select',
-        header: ({ table }) => {
+        header: ({table}) => {
           return (
             <div
               style={{
@@ -467,7 +473,7 @@ export const EmptyState: React.FC = () => {
           )
         },
         enableGrouping: false,
-        cell: ({ row }) => (
+        cell: ({row}) => (
           <div
             style={{
               display: 'flex',
@@ -488,7 +494,7 @@ export const EmptyState: React.FC = () => {
       },
       {
         id: 'name',
-        header: () => <div style={{ textAlign: 'center' }}>Name</div>,
+        header: () => <div style={{textAlign: 'center'}}>Name</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -511,7 +517,7 @@ export const EmptyState: React.FC = () => {
       },
       {
         id: 'otherInfo',
-        header: () => <div style={{ textAlign: 'center' }}>Other info</div>,
+        header: () => <div style={{textAlign: 'center'}}>Other info</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -628,8 +634,8 @@ export const Loading: React.FC = () => {
     columnResizeMode: 'onChange',
     manualSorting: false,
     initialSortBy: [
-      { id: 'firstName', desc: true },
-      { id: 'lastName', desc: false },
+      {id: 'firstName', desc: true},
+      {id: 'lastName', desc: false},
     ],
   }
 
@@ -637,7 +643,7 @@ export const Loading: React.FC = () => {
     () => [
       {
         id: 'select',
-        header: ({ table }) => {
+        header: ({table}) => {
           return (
             <div
               style={{
@@ -657,7 +663,7 @@ export const Loading: React.FC = () => {
           )
         },
         enableGrouping: false,
-        cell: ({ row }) => (
+        cell: ({row}) => (
           <div
             style={{
               display: 'flex',
@@ -678,7 +684,7 @@ export const Loading: React.FC = () => {
       },
       {
         id: 'name',
-        header: () => <div style={{ textAlign: 'center' }}>Name</div>,
+        header: () => <div style={{textAlign: 'center'}}>Name</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -701,7 +707,7 @@ export const Loading: React.FC = () => {
       },
       {
         id: 'otherInfo',
-        header: () => <div style={{ textAlign: 'center' }}>Other info</div>,
+        header: () => <div style={{textAlign: 'center'}}>Other info</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -834,9 +840,9 @@ export const EditableTemplateCell: React.FC = () => {
   }
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function
-  const onSorting = () => { }
+  const onSorting = () => {}
   // eslint-disable-next-line @typescript-eslint/no-empty-function
-  const onChangeRowSelection = () => { }
+  const onChangeRowSelection = () => {}
 
   const columns = React.useMemo<Array<TableColumnDef<Person>>>(
     () => [
@@ -1035,7 +1041,7 @@ const PersonTemplateCell = () => {
         md={2}
         lg={2}
         xl={2}
-        css={{ display: 'flex', justifyContent: 'center', gap: 5 }}
+        css={{display: 'flex', justifyContent: 'center', gap: 5}}
       >
         <Button onClick={handleConfirm} size='sm'>
           <Icon icon={faCheck} />
@@ -1080,8 +1086,8 @@ export const DataGrid: React.FC = () => {
     columnResizeMode: 'onChange',
     manualSorting: false,
     initialSortBy: [
-      { id: 'firstName', desc: true },
-      { id: 'lastName', desc: false },
+      {id: 'firstName', desc: true},
+      {id: 'lastName', desc: false},
     ],
   }
 
@@ -1096,7 +1102,7 @@ export const DataGrid: React.FC = () => {
     () => [
       {
         id: 'select',
-        header: ({ table }) => {
+        header: ({table}) => {
           return (
             <div
               style={{
@@ -1116,7 +1122,7 @@ export const DataGrid: React.FC = () => {
           )
         },
         enableGrouping: false,
-        cell: ({ row }) => (
+        cell: ({row}) => (
           <div
             style={{
               display: 'flex',
@@ -1137,7 +1143,7 @@ export const DataGrid: React.FC = () => {
       },
       {
         id: 'name',
-        header: () => <div style={{ textAlign: 'center' }}>Name</div>,
+        header: () => <div style={{textAlign: 'center'}}>Name</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -1160,7 +1166,7 @@ export const DataGrid: React.FC = () => {
       },
       {
         id: 'otherInfo',
-        header: () => <div style={{ textAlign: 'center' }}>Other info</div>,
+        header: () => <div style={{textAlign: 'center'}}>Other info</div>,
         footer: (props) => props.column.id,
         enableGrouping: false,
         columns: [
@@ -1253,6 +1259,28 @@ export const DataGrid: React.FC = () => {
           ></Pagination>
         </Table.Footer>
       </Table>
+    </div>
+  )
+}
+
+const StatusComponent = ({status}: {status: string}) => {
+  const renderIcon = () => {
+    switch (status) {
+      case 'relationship':
+        return <FontAwesomeIcon icon={faHeart} />
+      case 'complicated':
+        return <FontAwesomeIcon icon={faHeartBroken} />
+      case 'single':
+        return <FontAwesomeIcon icon={faUser} />
+      default:
+        return null
+    }
+  }
+
+  return (
+    <div>
+      <span style={{paddingRight: '5px'}}>{renderIcon()}</span>
+      <span>{status}</span>
     </div>
   )
 }
